requirements.txt
einops==0.7.0
gym==0.21.0
imageio==2.33.1
isaacgym.egg==info
matplotlib==3.7.4
moviepy==1.0.3
numpy==1.24.3
opencv_python==4.9.0.80
Pillow==10.2.0
piqa==1.3.2
pytorch_lightning==2.1.3
pyvirtualdisplay==3.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.12.0
stable_baselines3==1.5.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
wandb==0.16.2
